fix: crash/performance overlay copy buttons now use 3-tier clipboard fallback
The performance report panel and crash overlay create raw DOM elements
(via paintPerformancePrompt / paint), whose copy buttons only called
navigator.clipboard?.writeText() — which silently fails in Wails webview
(non-secure context, no Wails runtime fallback, no execCommand fallback).

Extract the 3-tier writeClipboardText (async API → Wails runtime →
textarea + execCommand) into src/lib/clipboard.ts and use it from both
CopyButton.tsx and crash.ts.

Also:
- Add type="button" to the overlay buttons (default submit type can
  interfere in some contexts)
- Add visual feedback: button text flips to "已复制"/"Copied" for
  1.2s on click (matching CopyButton behavior)
- Add dev-only __debugShowPerformancePanel / __debugShowCrashOverlay
  helpers for manual QA
